[feat] load select options with excel (#1873)
Se añade una opción de configuración en el campo SELECT para poder
cargar los valores y etiquetas desde excel, toda la operación se realiza
desde front, a backend solamente le tendría que llegar los arrays de
strings.

Si el select tiene labels no se puede editar. 

Si se "sube" un archivo con labels pasan a ser obligatorios, es decir,
cada value tiene que tener su label, sino es opcional.


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1206262666105108

---------

+++ server/src/api/public/__types.ts
@@ -3821,13 +3821,13 @@ export type PetitionUserPermission = PetitionPermission &
   };
 
 export type PetitionVariable = {
-  defaultValue: Scalars["Int"]["output"];
+  defaultValue: Scalars["Float"]["output"];
   name: Scalars["String"]["output"];
 };
 
 export type PetitionVariableResult = {
   name: Scalars["String"]["output"];
-  value: Maybe<Scalars["Int"]["output"]>;
+  value: Maybe<Scalars["Float"]["output"]>;
 };
 
 export type Profile = Timestamps & {
